chore: add a circleci config (#10)
## Which problem is this PR solving?

This adds a circleci config so that we can run smoke tests as part of
presubmits.

## Short description of the changes

This is a Frankenstein's monster of the circleci configs from ios and
android, but with the test calls replaced with calls to the existing
Makefile.

## How to verify that this has the expected result

This might take a few tries to get it right.

---

~- [ ] CHANGELOG is updated~ N/A
~- [ ] README is updated with documentation~ N/A