feat: Navigation instrumentation (#5)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

This adds some rudimentary instrumentation for Navigation in ReactNative
Router and Expo Router.

- Closes #<enter issue here>

## Short description of the changes

This adds a new component `<NavigationInstrumentation/>` Which can be
placed in the main layout or App file. When using react native
navigation it should be placed as a child component to
`<NavigationContainer/>`.

## How to verify that this has the expected result

This was manually tested but there is a todo in general for adding smoke
tests and automated tests in this repo.

---

- [X] CHANGELOG is updated
- [X] README is updated with documentation