feat: Add Release CI (#14)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?
- We had a manual release process to get through fedex week. Now we are
adding automated releases vis CircleCI.

- Closes #<enter issue here>

## Short description of the changes
Updates the CircleCI config and `RELEASING.md` files to automatically
release.

## How to verify that this has the expected result
We will be attempting a release later, if this works we will see it
published in NPM.

---

- [X] CHANGELOG is updated
- [X] README is updated with documentation