feat: add default instrumentation (#19)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

Our web and mobile sdks all provide autoinstrumentation that is enabled
by default with an option to disable them if desired.

- Closes #<enter issue here>

## Short description of the changes

Automatically adds instrumentation for `fetch` and uncaught exceptions
with the option for disabling in the config.

## How to verify that this has the expected result

Existing error tests should ensure this enables the instrumentation by
default.

---

- [X] CHANGELOG is updated
- [X] README is updated with documentation