feat: slow event loop instrumentation (#20)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

We did not have a system in place to auto-detect when the event loop may
be slow.

- Closes #<enter issue here>

## Short description of the changes

Adds autoinstrumentation that regularly polls and uses the detected
intervals to detect how long the event loop is taking to iterate. If it
detects a longer than normal time between polling it will emit a `slow
event loop` span. Both the polling interval and the stall threshold can
be configured.

## How to verify that this has the expected result

Automated tests have been added to ensure that the slow event loop
detection only sends a span on slow loop iterations.

---

- [X] CHANGELOG is updated
- [X] README is updated with documentation