feat: IOS Native modules (#23)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- Closes #<enter issue here>

## Short description of the changes

We include the Honeycomb OptenTelemetry swift library and initialize it
in an IOS module. A few things left to do before this PR can be merged:

- [x] Figure out configuration, The libraries need to be initialized at
app start before JS is loaded
- [x] Figure out integration patterns moving forward.

## How to verify that this has the expected result

---

- [ ] CHANGELOG is updated
- [ ] README is updated with documentation