maint: make lint/format work better (#30)
## Which problem is this PR solving?

Lint and auto-formatting weren't working for swift/objc/kotlin code, and
I was spending too much time futzing with formatting.

## Short description of the changes

1. Added `.editorconfig` for Android Studio to automatically format
correctly.
2. Added `.clang-format` for specifying (Objective) C(++) style.
3. Added `.swift-format` for specifying Swift style.
4. Added `spotless` config for linting Android code.
5. Updated `Makefile` to lint all the different kinds of code.
6. Ran `make format` and `make lint`.
7. Changed circleci config to run separate lint on android and swift, as
needed.

The config files are based on our Swift and Android repos.

Caveat: This doesn't run for the Android Turbo Module code -- only for
the example app. The build system was too weird for me to figure out how
to add it. Maybe we can add that in the future.

## How to verify that this has the expected result

I manually verified the changes, and all of the tests should still be
passing.

---

~- [ ] CHANGELOG is updated~ N/A
~- [ ] README is updated with documentation~ N/A