docs: Include build configuration in Native Modules docs (#34)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

When setting up a React Native application to use Native Modules, the
`Podfile` on iOS and the `build.gradle` file in android both need
additional configuration in order to build properly.

- Closes #<enter issue here>

## Short description of the changes

Updates the Readme to include steps to update the native build files.

## How to verify that this has the expected result

---

- [X] CHANGELOG is updated
- [X] README is updated with documentation