fix: add a default session id (#38)
## Which problem is this PR solving?

This PR makes it so that if the developer doesn't initialize the native
SDKs, then we still provide a reasonable default.

## Short description of the changes

The implementation is the same as the one on web. It's slightly
different than the native SDKs, in that the session id is not persisted,
so it's different on every run of the app. But this is still better than
the previous behavior of not providing any session id at all.

## How to verify that this has the expected result

I manually tested this by commenting out the native SDK initialization
in the example app and running the smoke tests.

I don't think it's worth trying to a test for this, because:
1. we want people to enable the native SDK
2. we would have to add a second example app just to test it

If we decide this is important functionality, we can always add a test
in the future.

---

- [x] CHANGELOG is updated
~- [ ] README is updated with documentation~ N/A