chore: remove github actions (#42)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?
React native creates the repo with github actions definitions by
default. We use circle for our CI so these are not needed.

- Closes #<enter issue here>

## Short description of the changes
Remove workflow files that are related to github actions.

## How to verify that this has the expected result
the react native github actions will not run on this pr

---

- [n/a] CHANGELOG is updated
- [n/a] README is updated with documentation