feat: Add expo plugin to generate source map UUID (#47)
## Which problem is this PR solving?
we need to link source maps to a specific build

## Short description of the changes
The plugin:
- generates a uuid
- writes it into the `manifest.xml` / `info.plist`
- writes that uuid into a file on the filesystem (in
`$cwd/dist/tmp/hny/SourceMapUuid.txt`
- future PRs/guidance will use this uuid when uploading the source maps,
similar to how we implemented proguard uploads on Android

This PR also adds code to our native bridges to read that value and
attach it as a resource attribute.

### Usage
Add the following line to the plugins section of
`app.json`/`app.config.ts`:

```js
{
  "expo": {
    "plugins": [
      // ...
      ["@honeycombio/opentelemetry-react-native"],
    ],
  }
}
```

## How to verify that this has the expected result
I added this library to our internal react native app and verified the
source map attribute appeared in our telemetry.

---

- [ ] CHANGELOG is updated
- [ ] README is updated with documentation