docs: fetch instrumentation (#51)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?
We needed to provide a link to the OpenTelemetry JS's fetch
instrumentation that we use.

- Closes #<enter issue here>

## Short description of the changes
Adds section for fetch instrumentation with a link to the documentation
for it.

## How to verify that this has the expected result

---

- [N/A] CHANGELOG is updated
- [X] README is updated with documentation