feat: add attributes for build enviornment (#52)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?
It is difficult to tell with JS exceptions if they are coming from a
development or a production build.

- Closes #<enter issue here>

## Short description of the changes
Add a resource attribute to all outgoing JS spans to indicate if this is
`development` or `production` builds.

## How to verify that this has the expected result

Tests have been added to ensure this

---

- [X] CHANGELOG is updated
- [X] README is updated with documentation