fix: remove broken plugin code that writes a text file (#63)
## Which problem is this PR solving?

The UUID plugin currently generates a text file at
`./dist/tmp/hny/SourceMapUuid.txt` whenever it's loaded. But that's not
how react plugins are supposed to work. The iOS and Android plugins only
generate new UUIDs during actual pre-builds. So, often the UUID in the
text file doesn't match the one in the code. Using it would cause
problems. And using the ones from iOS and Android directly isn't that
difficult.

## Short description of the changes

This just removes the text file generation from the plugin. Making the
plugin work correctly would be pretty painful.

## How to verify that this has the expected result

Tested manually.

---

- [X] CHANGELOG is updated
N/A ~- [ ] README is updated with documentation~